% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data22.R
\docType{data}
\name{data22}
\alias{data22}
\title{Tree Census Data}
\format{
## `data22`
A data frame with 30,808 rows and 18 columns:
\describe{
  \item{quad_sub_quad}{quadrat-subquadrat}
  \item{stem_tag}{StemTag, unique id}
  \item{species}{Species}
  \item{qx}{Cordinate x}
  \item{qy}{Cordinate y}
  \item{dbh}{Diameter at Breast Height}
  \item{status}{Status of Tree}
  \item{crown_position}{Position in Canopy}
  \item{percentage_of_crown_intact}{Percentage of Crown Intact}
  \item{percentage_of_crown_living}{Percentage of Crown Living}
  \item{lean_angle_if_greater_than_15_degrees}{Lean in Degrees if greater than 15 degrees}
  \item{fad}{Factors Associated with Death}
  \item{wounded_main_stem}{Wounded Main Stem}
  \item{canker_swelling_deformity}{Canker Swelling Deformity}
  \item{rotting_trunk}{Rotting Trunk}
  \item{lianas}{Lianas on Tree}
  \item{dead_with_resprout}{Dead with a Resprout}
  \item{notes}{Notes}

}
}
\source{
Dr. Dan J Johnson, University of Florida
}
\usage{
data22
}
\description{
Health data on trees in the Harvard Forest ForestGEO plot for the year 2022
}
\keyword{datasets}
