% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_means.R
\name{field_means}
\alias{field_means}
\title{Compute Means for each Numeric Tree Health Variable}
\usage{
field_means(data, x, y)
}
\arguments{
\item{data}{A data frame with numeric data variables from the tree dataset}

\item{x}{A string naming the numeric variable to take the mean from}

\item{y}{A string naming the grouping categorical variable}
}
\value{
A data frame a one row for each level of `y` (the categorical variable) containing the mean of `x` (the numeric variabl
}
\description{
`field_means()` calculates the mean of a numeric variable (like DBH,
percentage of crown living, etc.) grouped by a categorical variable (like the
species, crown position, or status).
}
\examples{
field_means(data21, x = "dbh", y = "species")

}
