% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_tree.R
\name{future_tree}
\alias{future_tree}
\title{Predicts Future Tree Survival Based on Past Health}
\usage{
future_tree(all_data, stem_tag, method = "mean_prob")
}
\arguments{
\item{all_data}{A data frame containing the complete tree dataset}

\item{stem_tag}{The individual tree ID}

\item{method}{statistical prediction method}
}
\value{
A numeric probability between 0-1 representing the likelihood of a specific tree's survival to the next year
}
\description{
`future_tree()` estimates the chances of a tree surviving into a
future year based on its observed health measurements across the 2021–2024
data set. The function uses probabilities to generate a prediction of survival status based on status and varying levels of external damage/sickness observed.
}
\examples{
# Generate the combined dataset
all_data <- comb_frame(data21, data22, data23, data24)
future_tree(all_data, stem_tag = 11769)
}
