% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_trees.R
\name{graph_trees}
\alias{graph_trees}
\title{Generating Visualizations for Measures of Tree Health, Mortality, and Biodiversity}
\usage{
graph_trees(
  data,
  graph_type = c("crown_position", "biodiv_mortality", "health_dist", "status_bar"),
  health_tree = "percentage_crown_living"
)
}
\arguments{
\item{data}{A data frame containing tree health metrics, either at the individual tree level or at the quadrat level depending on whether or not the output of the biodiv_trees() function is used.}

\item{graph_type}{A character string indicating which type of plot to generate.}

\item{health_tree}{Character string specifying which variable that constitutes an indice of tree health to plot, with the default variable in this function being percentage_crown_living.}
}
\value{
A ggplot2 graphics object that prints the graph/plot to the user screen.
}
\description{
Generates visualizations from the dataset for a given year between 2021-2024. Can include the apended quadrat-level biodiversity (output of biodiv_trees() function) if the user chooses. Automatically produces various user-friendly/easily-interpretable graphs based on the data that can be used to analyze tree health, mortality/status, and biodiversity. Users can select which type of graph which variables they would like to use to analyze the data.
}
\details{
This function is designed to help users more efficiently explore relationships between key variables like biodiversity, crown condition (percentage intact and percentage living, for example), and categorical variables such as mortality status and factors associated with death (fad). ggplot2 is required to run this function.
}
\examples{
graph_trees(data21, graph_type = "biodiv_mortality")
graph_trees(data23, graph_type = "status_bar")
}
