% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_species.R
\name{quad_species}
\alias{quad_species}
\title{Getting a List of Quadrats with the Most or Least of a Tree Species}
\usage{
quad_species(df, species, most = TRUE)
}
\arguments{
\item{df}{the data frame you want to search through}

\item{species}{the species you want to find}

\item{most}{defulat true, meaning you are getting quadrats with the most of that spieces}
}
\value{
list of top 5 quadrats
}
\description{
WRITE
}
\details{
WRITE
}
\examples{
quad_species(data21, "SP", most = TRUE)
quad_species(data22, "SP", most= FALSE)

}
