% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_frame.R
\name{comb_frame}
\alias{comb_frame}
\title{Combine the data frames from multiple years into a single data frame}
\usage{
comb_frame(data21, data22, data23, data24)
}
\arguments{
\item{data21}{A data frame containing the tree data frame for year 2021}

\item{data22}{A data frame containing the tree data frame for year 2022}

\item{data23}{A data frame containing the tree data frame for year 2023}

\item{data24}{A data frame containing the tree data frame for year 2024}
}
\value{
A data frame containing all cleaned and combined observations across the
four years, with a new column of the categorical variable `year`.
}
\description{
`comb_frame()` merges the data frames from each year in the package into one
data frame. It creates a seperate `year` categorical variable for the data across the years 2021-2024
}
\examples{
comb_frame(data21, data22, data23, data24)

}
