% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_tree.R
\name{future_tree}
\alias{future_tree}
\title{Predicts Future Tree Survival Based on Past Health}
\usage{
future_tree(data, stem_tag, method = "mean_prob")
}
\arguments{
\item{all_data}{A data frame containing the complete tree dataset}

\item{stem_tag}{The individual tree ID}

\item{method}{statistical prediction method}
}
\value{
A numeric probability between 0-1 representing the likelihood of a specific tree's survival to the next year
}
\description{
`future_tree()` estimates the chances of a tree surviving into a
future year based on its observed health measurements across the 2021–2024
dataset. The function uses probabilities to generate a prediction of survival status.
}
\examples{
future_tree(all_data, stem_tag = 174002)
}
