% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_health.R
\name{logistic_reg_health}
\alias{logistic_reg_health}
\title{Logistic Regression By Health Status}
\usage{
logistic_reg_health(df, x, ...)
}
\arguments{
\item{df}{the data frame}

\item{x}{the first variable in the logistic regression model}

\item{...}{additional variables in the logistic regression model}
}
\value{
The best logistic regression model.
}
\description{
NOT DONE
}
\details{
NOT DONE
}
\examples{
logistic_reg_health(data21, "species", "dbh")


}
