% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_species.R
\name{quad_species}
\alias{quad_species}
\title{Getting a List of Quadrats with the Most or Least of a Tree Species}
\usage{
quad_species(df, species, most = TRUE)
}
\arguments{
\item{df}{the data frame you want to search through}

\item{species}{the species you want to find}

\item{most}{default true, meaning you are getting quadrats with the most of that spieces, false means get the quadrats with the least of that species}
}
\value{
data frame of top 5 quadrats and counts of trees per quadrat
}
\description{
This function takes a data frame, species of tree and finds the quadrats with the most or least of this species.
}
\details{
Used to find quadrats with either the most or least of one species using counts. The data frame that is returned will contain the top 5 quadrats that meet this critera.
}
\examples{
quad_species(data21, "querru", most = TRUE)
quad_species(data22, "tsugca", most= FALSE)

}
